Correct some type for rtp value

+++ RTSP/Rtp/RawMediaFrame.cs
@@ -21,7 +21,7 @@ namespace Rtsp.Rtp
         }
 
         public required DateTime ClockTimestamp { get; init; }
-        public required ulong RtpTimestamp { get; init; }
+        public required uint RtpTimestamp { get; init; }
 
 
         public RawMediaFrame(IEnumerable<ReadOnlyMemory<byte>> data, IEnumerable<IMemoryOwner<byte>> owners)
